Upgrade dependencies for improved performance and compatibility
Updated "next" to 15.3.0, as well as several sharp and libvips-related libraries to their latest stable versions. These changes optimize dependency management and maintain consistency across environments.

Fix errors. Building - ok.
