+++ package.json
@@ -18,7 +18,7 @@
     "i18next": "^24.2.2",
     "i18next-browser-languagedetector": "^8.0.4",
     "js-cookie": "^3.0.5",
-    "next": "15.1.7",
+    "next": "15.3",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-i18next": "^15.4.1",
