+++ src/app/[locale]/admin/mainAdminStyles.module.scss
@@ -1,8 +1,8 @@
 .container{
   display: flex;
   flex-direction: column;
-  justify-content: start;
-  align-items: start;
+  justify-content: flex-start;
+  align-items: flex-start;
   width: 100%;
   height: 100%;
   padding: var(--spacing-perimeter-h) var(--spacing-perimeter-w);
