+++ src/app/[locale]/auth/login/Login.module.css
@@ -73,7 +73,7 @@
     gap: 10px;
     height: 75%;
     justify-self: center;
-    align-self: start;
+    align-self: flex-start;
     /*border: 1px solid #89cc1c;*/
 }
