+++ src/app/[locale]/auth/register/Register.module.scss
@@ -71,7 +71,7 @@
     gap: 10px;
     height: 75%;
     justify-self: center;
-    align-self: start;
+    align-self: flex-start;
     /*border: 1px solid #89cc1c;*/
 }
