+++ src/app/components/Catalog_components/GameItem/GameItem.module.scss
@@ -88,7 +88,7 @@
     padding: 1vh 1vw;
     //background-color: coral;
     flex-direction: column;
-    justify-content: end;
+    justify-content: flex-end;
     box-sizing: border-box;
     gap: 5px;
