+++ src/app/components/Profile/UserStats/GameCollector/GameCollector.module.scss
@@ -37,7 +37,7 @@
                 .gcItem {
                     display: flex;
                     flex-direction: column;
-                    align-items: start;
+                    align-items: flex-start;
                     color: var(--main_page_text);
 
                     .gcItemCount {
