+++ src/app/components/Profile/UserStats/UserStats.module.scss
@@ -1,64 +1,3 @@
-//
-//.statsContainer {
-//    display: flex;
-//    flex-direction: column;
-//    justify-content: space-between;
-//    background: var(--gray-800);
-//    width: 100%;
-//    box-sizing: border-box;
-//
-//    .statsContainerName {
-//        display: flex;
-//        justify-content: start;
-//        text-transform: uppercase;
-//        font-family: 'Oswald', sans-serif;
-//        font-weight: 300;
-//        font-size: 3rem;
-//        margin-bottom: 3vh;
-//        color: var(--main_page_text);
-//    }
-//
-//    .statsContainerMain {
-//        display: flex;
-//        align-items: start;
-//        gap: 3vw;
-//
-//        .statsContainerMainItem{
-//            display: flex;
-//            flex-direction: column;
-//            border: 1px solid #89cc1c;
-//        }
-//
-//        .leftSide{
-//            display: flex;
-//            flex-basis: 60%;
-//            gap: 3vw;
-//            flex-direction: column;
-//            justify-content: space-between;
-//            border: 1px solid green;
-//            box-sizing: border-box;
-//        }
-//
-//        .statsContainerMainFrList {
-//             display: flex;
-//             flex-direction: column;
-//             flex-basis: 40%;
-//             height: inherit;
-//             background-color: #222;
-//             overflow: auto;
-//             border: 1px solid #ff9100;
-//         }
-//    }
-//}
-//
-//.ItemContainerName {
-//    font-size: 2rem;
-//    font-weight: 400;
-//    text-transform: uppercase;
-//    color: var(--main_page_text);
-//    margin-bottom: 15px;
-//}
-
 .statsContainer {
     display: flex;
     flex-direction: column;
@@ -69,7 +8,7 @@
 
     .statsContainerName {
         display: flex;
-        justify-content: start;
+        justify-content: flex-start;
         text-transform: uppercase;
         font-family: 'Oswald', sans-serif;
         font-weight: 300;
@@ -80,7 +19,7 @@
 
     .statsContainerMain {
         display: flex;
-        align-items: start;
+        align-items: flex-start;
         gap: 3vw;
 
         .statsContainerMainFlItem {
