+++ src/app/components/ui/EmailConfirmationNotice/EmailConfirmationNotice.module.scss
@@ -26,7 +26,7 @@
     display: flex;
     flex: 1;
     flex-direction: column;
-    align-items: start;
+    align-items: flex-start;
     justify-content: center;
   }
 }
