+++ src/app/page.module.css
@@ -13,7 +13,7 @@
   min-height: 100svh;
   padding: 80px;
   gap: 64px;
-  font-family: var(--font-geist-sans);
+  /*font-family: var(--font-geist-sans);*/
 }
 
 @media (prefers-color-scheme: dark) {
@@ -35,7 +35,7 @@
 }
 
 .main ol {
-  font-family: var(--font-geist-mono);
+  /*font-family: var(--font-geist-mono);*/
   padding-left: 0;
   margin: 0;
   font-size: 14px;
@@ -66,7 +66,6 @@
   border-radius: 128px;
   height: 48px;
   padding: 0 20px;
-  border: none;
   border: 1px solid transparent;
   transition:
     background 0.2s,
@@ -129,7 +128,6 @@ a.secondary {
 @media (max-width: 600px) {
   .page {
     padding: 32px;
-    padding-bottom: 80px;
   }
 
   .main {
